initial commit: Based on the tour of heroes tut
This is the initial version. It's a modified version of the result
of the Angular Tour of Heroes tutorial. A few bits and pieces differ
from the tutorial due to the different goals of this app and different
opinions held by the author.

This is the author's first frontend work since 2017, and their second
frontend work ever.